cleaned up seeding users script. separated users and automated output status
