+++ scripts/usersArray.js
@@ -1,0 +1,150 @@
+// seed users for testing
+
+const usersArray = [
+  {
+    "name" : "Mark",
+    "email" : "mark@hotmail.com",
+    "password" : "pass",
+
+    "phone" : "916-881-3553",
+    "gender" : "male",
+    "birthdate" : new Date(1970, 09, 23),
+    "image" : null,
+    "bio": "Hipster-friendly entrepreneur. Food expert. Gamer. Bacon guru. Music specialist. Beer enthusiast.",
+
+    "isActive" : true,
+    "recentLocation" : { latitude: 35.8069093, longitude: -78.818777 },
+    "acceptedChats" : [],
+    "ignoredChats" : [],
+    "pendingChats" : [],
+    "requestedChats" : []
+  },
+  {
+    "name" : "Merle",
+    "email" : "merle@hotmail.com",
+    "password" : "pass",
+
+    "phone" : "514-832-3563",
+    "gender" : "female",
+    "birthdate" : new Date(1990, 09, 18),
+    "image" : null,
+    "bio": "Extreme beer buff. Amateur social mediaholic. Alcohol fan. Hardcore travel advocate.",
+
+    "isActive" : true,
+    "recentLocation" : { latitude: 35.7849302, longitude: -78.6602814 },
+    "acceptedChats" : [],
+    "ignoredChats" : [],
+    "pendingChats" : [],
+    "requestedChats" : []
+  },
+  {
+    "name" : "Juana",
+    "email" : "juana@hotmail.com",
+    "password" : "pass",
+
+    "phone" : "732-635-4686",
+    "gender" : "female",
+    "birthdate" : new Date(1973, 08, 08),
+    "image" : null,
+    "bio": "Avid beer guru. Organizer. Typical explorer. Problem solver. Unapologetic analyst. Pop culture practitioner.",
+
+    "isActive" : true,
+    "recentLocation" : { latitude: 35.8069093, longitude: -78.818777 },
+    "acceptedChats" : [],
+    "ignoredChats" : [],
+    "pendingChats" : [],
+    "requestedChats" : []
+  },
+  {
+    "name" : "Vernon",
+    "email" : "vernon@hotmail.com",
+    "password" : "pass",
+
+    "phone" : "602-660-0766",
+    "gender" : "male",
+    "birthdate" : new Date(1990, 07, 03),
+    "image" : null,
+    "bio": "Award-winning tvaholic. Typical coffee enthusiast. Pop culture junkie. Lifelong alcohol fanatic. Food practitioner. Avid web geek.",
+
+    "isActive" : true,
+    "recentLocation" : { latitude: 35.8069093, longitude: -78.818777 },
+    "acceptedChats" : [],
+    "ignoredChats" : [],
+    "pendingChats" : [],
+    "requestedChats" : []
+  },
+  {
+    "name" : "Lisa",
+    "email" : "lisa@hotmail.com",
+    "password" : "pass",
+
+    "phone" : "225-447-1703",
+    "gender" : "female",
+    "birthdate" : new Date(1950, 10, 08),
+    "image" : null,
+    "bio": "Prone to fits of apathy. Devoted music geek. Troublemaker. Typical analyst. Alcohol practitioner. Food junkie. Passionate tv fan. Web expert.",
+
+    "isActive" : true,
+    "recentLocation" : { latitude: 35.8069093, longitude: -78.818777 },
+    "acceptedChats" : [],
+    "ignoredChats" : [],
+    "pendingChats" : [],
+    "requestedChats" : []
+  },
+  {
+    "name" : "Harold",
+    "email" : "harold@hotmail.com",
+    "password" : "pass",
+
+    "phone" : "206-369-3139",
+    "gender" : "male",
+    "birthdate" : new Date(1993, 01, 12),
+    "image" : null,
+    "bio": "Freelance bacon junkie. Wannabe internet specialist. Music aficionado. Lifelong twitter maven.",
+
+    "isActive" : true,
+    "recentLocation" : { latitude: 35.8069093, longitude: -78.818777 },
+    "acceptedChats" : [],
+    "ignoredChats" : [],
+    "pendingChats" : [],
+    "requestedChats" : []
+  },
+  {
+    "name" : "Jeremy",
+    "email" : "jeremy@hotmail.com",
+    "password" : "pass",
+
+    "phone" : "816-202-2403",
+    "gender" : "male",
+    "birthdate" : new Date(1971, 01, 01),
+    "image" : null,
+    "bio": "Professional communicator. Certified food fanatic. Extreme internet practitioner. Music fan.",
+
+    "isActive" : true,
+    "recentLocation" : { latitude: 35.7849302, longitude: -78.6602814 },
+    "acceptedChats" : [],
+    "ignoredChats" : [],
+    "pendingChats" : [],
+    "requestedChats" : []
+  },
+  {
+    "name" : "Juan",
+    "email" : "juan@hotmail.com",
+    "password" : "pass",
+
+    "phone" : "317-234-2198",
+    "gender" : "male",
+    "birthdate" : new Date(1976, 06, 10),
+    "image" : null,
+    "bio": "Analyst. Freelance alcoholaholic. Hardcore coffee fanatic. Food scholar. Music junkie. Beer ninja. Problem solver. Pop culture evangelist.",
+
+    "isActive" : true,
+    "recentLocation" : { latitude: 35.8069093, longitude: -78.818777 },
+    "acceptedChats" : [],
+    "ignoredChats" : [],
+    "pendingChats" : [],
+    "requestedChats" : []
+  },
+];
+
+module.exports = usersArray;
